+++ run/kubejs/server_scripts/data/food.js
@@ -3,9 +3,7 @@ TFCEvents.data(e => {
         ingredient: 'minecraft:bread',
         food: {
             hunger: 4,
-            nutrients: [
-                0, 0, 1, 0, 0
-            ]
+            fruit: -20
         }
     });
 })