+++ run/kubejs/server_scripts/worldgen/world.js
@@ -1,0 +1,11 @@
+
+ServerEvents.registry('tfc:worldgen/rock_settings', event => {
+    event.create('vanilla')
+        .raw('minecraft:stone')
+        .hardened('minecraft:deepslate')
+        .gravel('minecraft:gravel')
+        .cobble('minecraft:cobblestone')
+        .sand('minecraft:sand')
+        .sandstone('minecraft:sandstone')
+        .mafic()
+})