+++ run/kubejs/startup_scripts/events/world.js
@@ -5,13 +5,5 @@ TFCEvents.defaultWorldSettings(e => {
     e.setTemperatureScale(10000);
     e.setRainfallScale(10000);
 
-    e.addRock('test', {
-        raw: 'minecraft:stone',
-        hardened: 'minecraft:deepslate',
-        gravel: 'minecraft:gravel',
-        cobble: 'minecraft:cobblestone',
-        sand: 'minecraft:sand',
-        sandstone: 'minecraft:sandstone',
-        mafic: true
-    }, true);
+    e.addRock('test', 'kubejs:vanilla', true);
 })