+++ src/main/java/io/github/notenoughmail/kubejstfc/util/commands/impl/WorldPrinter.java
@@ -1,6 +1,8 @@
 package io.github.notenoughmail.kubejstfc.util.commands.impl;
 
 import com.mojang.brigadier.context.CommandContext;
+import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import com.mojang.brigadier.exceptions.DynamicCommandExceptionType;
 import io.github.notenoughmail.kubejstfc.util.Assistant;
 import io.github.notenoughmail.kubejstfc.util.Printer;
 import io.github.notenoughmail.kubejstfc.util.commands.KubeJSTFCCommands;
@@ -8,10 +10,13 @@ import net.dries007.tfc.network.ChunkWatchPacket;
 import net.dries007.tfc.world.ChunkGeneratorExtension;
 import net.dries007.tfc.world.chunkdata.ChunkData;
 import net.dries007.tfc.world.settings.RockLayerSettings;
+import net.dries007.tfc.world.settings.RockSettings;
 import net.dries007.tfc.world.settings.Settings;
 import net.minecraft.ChatFormatting;
 import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.commands.arguments.ResourceKeyArgument;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Holder;
 import net.minecraft.network.chat.ClickEvent;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.HoverEvent;
@@ -19,7 +24,12 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.phys.Vec3;
 
+import java.util.Map;
+import java.util.Objects;
+
 public interface WorldPrinter {
+
+    DynamicCommandExceptionType ERROR_INVALID_ROCK_SETTING = new DynamicCommandExceptionType(o -> Component.literal("There is no feature with type \"%s\"".formatted(o)));
 
     static int worldSettings(CommandContext<CommandSourceStack> ctx) {
         if (ctx.getSource().getLevel().getChunkSource().getGenerator() instanceof ChunkGeneratorExtension ext) {
@@ -61,8 +71,20 @@ public interface WorldPrinter {
                                             .appendMap(d.layers()),
                                     $ -> p.listItem()
                             ))
+                            .newLine()
                             .descriptor("rocks")
-                            .appendMap(rockData.rocks(), (p, h) -> p.appendRecordAsMap(h.value()))
+                            .appendMap(rockData.rocks(), (p, h) -> {
+                                if (h.kind() == Holder.Kind.REFERENCE) {
+                                    final String id = Objects.requireNonNull(h.getKey()).location().toString();
+                                    p.append(Component.literal(id).withStyle(s -> s
+                                            .withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/kubejs_tfc print_rock_setting %s".formatted(id)))
+                                            .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal("Print %s".formatted(id))))
+                                            .withColor(ChatFormatting.AQUA)
+                                    ));
+                                } else {
+                                    p.appendRecordAsMap(h.value());
+                                }
+                            })
                             .getFormattedText(),
                     ctx
             );
@@ -70,6 +92,20 @@ public interface WorldPrinter {
         } else {
             return KubeJSTFCCommands.failMsg("Not a TFC-like level!", ctx);
         }
+    }
+
+    static int rockSetting(CommandContext<CommandSourceStack> ctx) throws CommandSyntaxException {
+        final Holder.Reference<RockSettings> rock = ResourceKeyArgument.resolveKey(ctx, "rock", RockSettings.KEY, ERROR_INVALID_ROCK_SETTING);
+        final Printer printer = Printer.create()
+                .appendPlain("Rock setting '%s' is:\n".formatted(rock.key().location()));
+        final Map<String, ?> fields = Printer.convertRecordToMap(rock.value());
+        Assistant.iterate(
+                fields.entrySet(),
+                e -> printer.append(e.getKey(), e.getValue(), true),
+                $ -> printer.newLine()
+        );
+        KubeJSTFCCommands.sysMsg(printer.getFormattedText(), ctx);
+        return 1;
     }
 
     static int chunkData(CommandContext<CommandSourceStack> ctx) {